requirements.txt
nextcord.py
humanfriendly
pillow
python-dotenv

